[Fix] #94 - 취득한 자격증 api 통합
