+++ CERTI-iOS/Network/Acquisition/AcquisitionService.swift
@@ -11,6 +11,7 @@ import Moya
 
 protocol AcquisitionServiceProtocol {
     func fetchAcquisitionList() async -> Result<AcquisitionListResponseDTO, NetworkError>
+    func addAcquisition(certificationId: Int) async -> Result<BaseResponseDTO<Bool>, NetworkError>
 }
 
 final class AcquisitionService: BaseService, AcquisitionServiceProtocol {
@@ -18,5 +19,9 @@ final class AcquisitionService: BaseService, AcquisitionServiceProtocol {
     
     func fetchAcquisitionList() async -> Result<AcquisitionListResponseDTO, NetworkError> {
         return await requestDecodable(provider, .fetchAcquisitionList)
+    }
+    
+    func addAcquisition(certificationId: Int) async -> Result<BaseResponseDTO<Bool>, NetworkError> {
+        return await requestDecodable(provider, .addAcquisition(certificationId: certificationId))
     }
 }
